+++ tests/Api/FitbitTest.php
@@ -213,8 +213,8 @@ class FitbitTest extends TestCase
         $this->assertTrue($this->fitbit->devices() instanceof \Namelivia\Fitbit\Api\Devices);
     }
 
-    public function testGettingAWeightInstance()
+    public function testGettingABodyInstance()
     {
-        $this->assertTrue($this->fitbit->weight() instanceof \Namelivia\Fitbit\Api\Weight);
+        $this->assertTrue($this->fitbit->weight() instanceof \Namelivia\Fitbit\Api\Body);
     }
 }
